[change] enum_bit 入出力をusizeに変更。

+++ src/algo/enum_bit/gray_code.rs
@@ -3,7 +3,7 @@
 //! <https://ja.wikipedia.org/wiki/%E3%82%B0%E3%83%AC%E3%82%A4%E3%82%B3%E3%83%BC%E3%83%89>
 
 /// `n`桁のGray codeを昇順に列挙する。
-pub fn gray_code(n: u32) -> impl Iterator<Item = u32> {
+pub fn gray_code(n: u32) -> impl Iterator<Item = usize> {
     (0..1 << n).map(|i| i ^ (i >> 1))
 }
