+++ src/algo/enum_bit/superset_asc.rs
@@ -2,7 +2,7 @@
 use std::iter::successors;
 
 /// ビット列としての`a`の上位集合を昇順に列挙するイテレータを返す。
-pub fn super_asc(a: u32, n: u32) -> impl Iterator<Item = u32> {
+pub fn super_asc(a: usize, n: u32) -> impl Iterator<Item = usize> {
     successors(Some(a), move |&t| {
         let t = (t + 1) | a;
         (t < 1 << n).then_some(t)
@@ -21,7 +21,7 @@ mod tests {
     #[test_case(0b10000000, 8)]
     #[test_case(0b10000001, 8)]
     #[test_case(0b11011011, 8)]
-    fn check(x: u32, n: u32) {
+    fn check(x: usize, n: u32) {
         let a = (0..1 << n).filter(|i| (x & !i) == 0).collect::<Vec<_>>();
 
         let b = super_asc(x, n).collect::<Vec<_>>();
