+++ src/algo/enum_bit/superset_desc.rs
@@ -2,7 +2,7 @@
 use std::iter::successors;
 
 /// ビット列としての`a`の上位集合を降順に列挙するイテレータを返す。
-pub fn superset_desc(a: u32, n: u32) -> impl Iterator<Item = u32> {
+pub fn superset_desc(a: usize, n: u32) -> impl Iterator<Item = usize> {
     let x = (1 << n) - 1;
     let y = x ^ (a & x);
 
@@ -21,7 +21,7 @@ mod tests {
     #[test_case(0b10000000, 8)]
     #[test_case(0b10000001, 8)]
     #[test_case(0b11011011, 8)]
-    fn check(x: u32, n: u32) {
+    fn check(x: usize, n: u32) {
         let a = (0..1 << n)
             .rev()
             .filter(|i| (x & !i) == 0)
